requirements.txt
absl-py==1.0.0
aniso8601==8.0.0
appdirs==1.4.4
asgiref==3.2.10
astunparse==1.6.3
attrs==20.2.0
boto3==1.17.14
botocore==1.20.112
cached-property==1.5.2
cachetools==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
drf-yasg==1.17.1
elasticsearch==7.8.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==2.10
image==1.5.33
importlib-metadata==4.11.3
inflection==0.5.1
install==1.3.5
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==1.1.1
numpy==1.21.5
oauthlib==3.2.0
openapi-codec==1.3.2
opencv-contrib-python==4.5.3.56
opt-einsum==3.3.0
packaging==20.4
Pillow==9.0.1
protobuf==3.19.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.2
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.7
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
swagger-spec-validator==2.7.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
typing_extensions==4.1.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.25
Werkzeug==1.0.1
wrapt==1.14.0
zipp==3.1.0
